requirements.txt
requests>=2.1.0
Pillow>=2.3.0